testDevops/requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-rest-knox==4.1.0
django-utils-six==2.0
Django-Verify-Email==1.0.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
promise==2.3
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-utils==2.5.6
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
utils==1.0.1
